requirements.txt
streamlit
pandas
streamlit-option-menu
google-genai
python-dotenv